+++ src/Contracts/HasName.php
@@ -1,0 +1,21 @@
+<?php
+
+namespace Digitlimit\Alert\Contracts;
+
+interface HasName
+{
+    /**
+     * Set the field name.
+     */
+    public function name(string $name): self;
+
+    /**
+     * Get the field name.
+     */
+    public function getName(): string;
+
+    /*
+     * Check if the field has a name.
+     */
+    public function hasName(): bool;
+}
