+++ src/Traits/WithName.php
@@ -1,0 +1,34 @@
+<?php
+
+namespace Digitlimit\Alert\Traits;
+
+trait WithName
+{
+    protected string $name;
+
+    /**
+     * Set the field name.
+     */
+    public function name(string $name): self
+    {
+        $this->name = $name;
+
+        return $this;
+    }
+
+    /**
+     * Get the field name.
+     */
+    public function getName(): string
+    {
+        return $this->name;
+    }
+
+    /*
+    * Check if the field has a name.
+    */
+    public function hasName(): bool
+    {
+        return isset($this->name);
+    }
+}
